# [process] add new fastpath to `DataScrubber`